--- src/localization/languages.ts
@@ -1,5 +1,0 @@
-export enum Languages {
-  english = 'en',
-  ukrainian = 'ua',
-  russian = 'ru',
-}
