+++ src/localization/localization.ts
@@ -1,10 +1,17 @@
 import i18n, {Callback, InitOptions} from 'i18next';
 import {initReactI18next} from 'react-i18next';
 
-import {Languages} from './languages';
+import {storeApi} from '../services/storage/storeApi';
+import {AsyncKey} from '../services/storage/asyncKey';
 import en from './translations/en.json';
 import ua from './translations/ua.json';
 import ru from './translations/ru.json';
+
+export enum Languages {
+  english = 'en',
+  ukrainian = 'ua',
+  russian = 'ru',
+}
 
 class Localization {
   public changeLanguage = async (key: Languages) =>
@@ -18,14 +25,23 @@ class Localization {
   public initialization = async (options: InitOptions, callback?: Callback) => {
     await i18n.use(initReactI18next).init({...options}, callback);
   };
-}
 
-export const InitLanguage = (language: Languages) =>
-  localization.initialization({
-    resources: {en, ua, ru},
-    lng: language || Languages.english,
-    keySeparator: false,
-    interpolation: {escapeValue: false},
-  });
+  public InitLanguage = async (language: Languages) =>
+    await localization.initialization({
+      resources: {en, ua, ru},
+      lng: language || Languages.english,
+      keySeparator: false,
+      interpolation: {escapeValue: false},
+    });
+
+  public saveLanguage = async (language: Languages) => {
+    await storeApi.set(AsyncKey.currentLanguage, language);
+  };
+
+  public getLanguage = async () => {
+    const currentLanguage = await storeApi.getString(AsyncKey.currentLanguage);
+    return currentLanguage as Languages;
+  };
+}
 
 export const localization = new Localization();
